:truck: move device related stuff into device module
