+++ sdk/device/configuration/base_config.py
