--- sdk/utils/configuration/__init__.py
@@ -1,10 +1,0 @@
-from sdk.utils.configuration.base_config import BaseConfig, InvalidConfiguration
-from sdk.utils.configuration.emu_free import EmuFreeConfig
-from sdk.utils.configuration.emu_sv import EmuSVConfig
-
-__all__ = [
-    "BaseConfig",
-    "InvalidConfiguration",
-    "EmuFreeConfig",
-    "EmuSVConfig",
-]
